+++ crates/openapi/src/routes/merchant_connector_account.rs
@@ -61,9 +61,9 @@
 )]
 pub async fn connector_create() {}
 
-/// Merchant Connector - Create
+/// Connector Account - Create
 ///
-/// Creates a new Merchant Connector for the merchant account. The connector could be a payment processor/facilitator/acquirer or a provider of specialized services like Fraud/Accounting etc.
+/// Creates a new Connector Account for the merchant account. The connector could be a payment processor/facilitator/acquirer or a provider of specialized services like Fraud/Accounting etc.
 #[cfg(feature = "v2")]
 #[utoipa::path(
     post,
@@ -146,7 +146,7 @@ pub async fn connector_create() {}
 )]
 pub async fn connector_retrieve() {}
 
-/// Merchant Connector - Retrieve
+/// Connector Account - Retrieve
 ///
 /// Retrieves details of a Connector account
 #[cfg(feature = "v2")]
@@ -235,9 +235,9 @@ pub async fn connector_list() {}
 )]
 pub async fn connector_update() {}
 
-/// Merchant Connector - Update
+/// Connector Account - Update
 ///
-/// To update an existing Merchant Connector account. Helpful in enabling/disabling different payment methods and other settings for the connector
+/// To update an existing Connector account. Helpful in enabling/disabling different payment methods and other settings for the connector
 #[cfg(feature = "v2")]
 #[utoipa::path(
     put,