+++ crates/router/src/services/connector_integration_interface.rs
@@ -307,6 +307,19 @@ impl api::IncomingWebhook for ConnectorEnum {
             Self::New(connector) => connector.get_external_authentication_details(request),
         }
     }
+
+    fn get_mandate_details(
+        &self,
+        request: &IncomingWebhookRequestDetails<'_>,
+    ) -> CustomResult<
+        Option<hyperswitch_domain_models::router_flow_types::ConnectorMandateDetails>,
+        errors::ConnectorError,
+    > {
+        match self {
+            Self::Old(connector) => connector.get_mandate_details(request),
+            Self::New(connector) => connector.get_mandate_details(request),
+        }
+    }
 }
 
 impl api::ConnectorTransactionId for ConnectorEnum {